requirements.txt
streamlit>=1.26.0
Pillow>=10.0.0
streamlit-js-eval==0.1.7 # fixme == ??
python-resize-image==1.1.20
streamlit_cropper>=0.3.1
